+++ server/src/server.ts
@@ -19,6 +19,7 @@ import * as referencesProvider from "./references-provider";
 import * as foldingRangeProvider from "./folding-range-provider";
 import * as documentHighlightProvider from "./document-highlight-provider";
 import * as renameProvider from "./rename-provider";
+import * as inlayHintProvider from "./inlay-hint-provider";
 import * as linter from "./linter";
 
 // Create a connection for the server (Node-IPC transport).
@@ -62,6 +63,7 @@ connection.onInitialize((params: InitializeParams) => {
       renameProvider: {
         prepareProvider: true,
       },
+      inlayHintProvider: true,
     },
   };
   if (hasWorkspaceFolderCapability) {
@@ -171,5 +173,7 @@ connection.onFoldingRanges(foldingRangeProvider.resolve);
 connection.onPrepareRename(renameProvider.prepare);
 
 connection.onRenameRequest(renameProvider.resolve);
+
+connection.languages.inlayHint.on(inlayHintProvider.resolve);
 
 connection.listen();